+++ system_theme/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * feat: Reactive theming for macOS ([#45](https://github.com/bdlukaa/system_theme/pull/45))
 * chore: Migrate iOS and macos to Swift Package Manager.
+* feat: Automatically adjust lightness if the platform doesn't support it natively.
+  This is enabled by default. You can disable it by setting `SystemTheme.autoAdjustLightness` to `false`.
 
 ## [3.1.2] - [04/10/2024]